requirements.txt
# RENTA Core Dependencies
# Generated from pyproject.toml [project.dependencies]
# For development dependencies, use: pip install -e .[dev]
# For scraping dependencies, use: pip install -e .[scraping]

# Core data processing
pandas==2.2.3
numpy==1.26.4

# Web scraping and HTTP
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.3.0

# Geospatial analysis
scikit-learn==1.5.2

# Configuration and validation
pyyaml==6.0.2
jsonschema==4.23.0

# AWS integration
boto3==1.35.70
botocore==1.35.70

# Template rendering
jinja2==3.1.4

# Logging and monitoring
structlog==24.4.0

# Progress bars
tqdm==4.67.1
